Project III/1215/requirements.txt
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
black==22.10.0
bleach==4.1.0
boto3==1.26.13
botocore==1.29.13
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
cssbeautifier==1.14.7
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-cleanup==6.0.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-froala-editor==4.0.15
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-shortcuts==1.6
django-social-share==2.3.0
django-storages==1.13.1
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-serializer-cache==0.3.4
EditorConfig==0.12.3
executing==1.2.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
ipython==8.6.0
jedi==0.18.1
jmespath==1.0.1
jsbeautifier==1.14.7
matplotlib-inline==0.1.6
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pathspec==0.10.1
phonenumbers==8.13.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.1
tomli==2.0.1
tqdm==4.64.1
traitlets==5.5.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
zipp==3.10.0
zope.interface==5.5.2
